release uploadfile intoverflow fix

+++ src/Picturepark.SDK.V1/Partial/TransferClient.cs
@@ -177,9 +177,12 @@ namespace Picturepark.SDK.V1
                         using (var fileStream = File.OpenRead(fileLocation.AbsoluteSourcePath))
                         {
                             var currentChunkSize = chunkSize;
-                            var position = (number - 1) * chunkSize;
+                            var position = (number - 1) * (long)chunkSize;
 
                             // last chunk may have a different size.
+                            // not using long because
+                            // - fileStream.Read's count argument is int
+                            // - position for the last chunk is already close to the end of the file
                             if (number == totalChunks)
                             {
                                 currentChunkSize = (int)(fileSize - position);
